requirements.txt
nicegui
smbus2
python-periphery
